move App stories to Home
